Second task is complete. Wacth description
There are 2 algorithms:
for L1 we are just finding the "median" of heatmap
for d (some number) we are summing up numbers in the columns near our number ( d steps away from it ) and getting the weight of the column that we are looking at. after that we finding the biggest number and the column, in which this number is, is the column that we should choose.