Second task is completed
The aim of this task was to implement both easy and smart way of summing up some sub-arrays. The results showed that easy way (using numpy) is much more effective for some small amount of counts of the sum (around 10^3 for 10^6 size of 1-d array). But if bigger amount of counts needed, then it is much better to use smart sum algorithms, which alows us, after some precalculations, to sum up sub-arrays for constant time.